+++ package.json
@@ -2,12 +2,12 @@
   "name": "honeyclient",
   "version": "1.0.0",
   "description": "Interacts with a malicious website to grab screenshots and other information.",
-  "main": "src/index.js",
+  "main": "src/index.mjs",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --experimental-modules src/index.js",
-    "server": "nodemon --exec node --experimental-modules src/index.js"
+    "start": "node --experimental-modules src/index.mjs",
+    "dev": "nodemon --exec node --experimental-modules src/index.mjs"
   },
   "repository": {
     "type": "git",
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@limedocs/express-middleware-memfs": "^1.0.0-beta.18",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "memfs": "^3.0.4",
     "puppeteer": "^1.20.0"
